Doing major refactoring on code usability
