+++ aioaquarea/errors.py
@@ -59,6 +59,7 @@ class AuthenticationErrorCodes(StrEnum):
     INVALID_USERNAME_OR_PASSWORD = "1001-1401"
     INVALID_CREDENTIALS = "1000-1401"
     API_ERROR = "API_ERROR"
+    TOKEN_EXPIRED = "TOKEN_EXPIRED" # Added for token expiration
 
 class DataNotAvailableError(Exception):
     """Exception raised when data is not available"""
