+++ main.py
@@ -8,13 +8,14 @@ import aiohttp
 import asyncio
 import logging
 from datetime import timedelta
+import os
 
 async def main():
     print("Starting Aquarea client...")
     async with aiohttp.ClientSession() as session:
         client = Client(
-            username="username@gmail.com",
-            password="your_password",
+            username=os.environ["USER_NAME"],
+            password=os.environ["PASSWORD"],
             session=session,
             device_direct=True,
             refresh_login=True,
@@ -22,9 +23,7 @@ async def main():
         )
 
         # The library is designed to retrieve a device object and interact with it:
-        print("Fetching devices...")
         devices = await client.get_devices()
-        print(f"Devices fetched: {devices}")
 
         # Picking the first device associated with the account:
         device_info = devices[0]
@@ -33,13 +32,8 @@ async def main():
             device_info=device_info, consumption_refresh_interval=timedelta(minutes=1)
         )
 
-        # Or the device can also be retrieved by its long id if we know it:
-        #device = await client.get_device(
-        #    device_id="LONG ID", consumption_refresh_interval=timedelta(minutes=1)
-        #)
-
         # Then we can interact with the device:
-        #await device.set_mode(UpdateOperationMode.HEAT)
+        await device.set_temperature(17, 1)
 
         # The device can automatically refresh its data:
         #await device.refresh_data()
